Mac OS X build files and compilation modifications

+++ java/exe/jpwr_rt_gdh/src/jpwr_rt_hist.cpp
@@ -477,7 +477,7 @@ int compareStr(char *ev, char *usr)
 {
   int sts;
   int startPos = 0;
-  unsigned int endPos = 0;
+  unsigned long endPos = 0;
   char *str1;
   sts = dcli_toupper(usr,usr);
   if(ODD(sts))
